Move the atom module into a separate hyperon-atom crate
