+++ lib/Cargo.toml
@@ -11,7 +11,6 @@ directories = "5.0.1" # For Environment to find platform-specific config locatio
 smallvec = "1.10.0"
 im = "15.1.0"
 rand = "0.9.0"
-bitset = "0.1.2"
 dyn-fmt = "0.4.0"
 itertools = "0.13.0"
 unescaper = "0.1.5"
@@ -26,6 +25,7 @@ semver = { version="1.0", features = ["serde"], optional=true }
 git2 = { version="0.18.3", features=["vendored-libgit2"], optional=true }
 
 hyperon-common = { workspace = true }
+hyperon-atom = { workspace = true }
 
 [dev-dependencies]
 ra_ap_profile = "0.0.261"
