+++ lib/src/lib.rs
@@ -1,5 +1,2 @@
-pub mod atom;
 pub mod space;
 pub mod metta;
-
-pub use atom::*;
