+++ lib/src/metta/interpreter.rs
@@ -1,8 +1,8 @@
 //! MeTTa assembly language implementation. See
 //! [minimal MeTTa documentation](https://github.com/trueagi-io/hyperon-experimental/blob/main/docs/minimal-metta.md) for details.
 
-use crate::*;
-use crate::atom::matcher::*;
+use hyperon_atom::*;
+use hyperon_atom::matcher::*;
 use crate::space::*;
 use crate::metta::*;
 use crate::metta::types::*;
@@ -345,7 +345,7 @@ impl FromIterator<VariableAtom> for Variables {
 }
 
 impl VariableSet for Variables {
-    type Iter<'a> = im::hashset::Iter<'a, atom::VariableAtom> where Self: 'a;
+    type Iter<'a> = im::hashset::Iter<'a, hyperon_atom::VariableAtom> where Self: 'a;
 
     fn contains(&self, var: &VariableAtom) -> bool {
         self.0.contains(var)
@@ -1448,7 +1448,7 @@ fn metta_call_return(args: Atom, bindings: Bindings) -> MettaResult {
 mod tests {
     use super::*;
     use crate::metta::text::metta_atom;
-    use crate::metta::space::grounding::metta_space;
+    use crate::space::grounding::metta_space;
     use hyperon_common::assert_eq_no_order;
 
     #[test]
