+++ lib/src/metta/runner/builtin_mods/fileio.rs
@@ -1,6 +1,6 @@
 use std::cell::RefCell;
 use std::fmt::{Display, Formatter};
-use crate::{sym, Atom, ExecError, Grounded};
+use hyperon_atom::{sym, Atom, ExecError, Grounded};
 use crate::metta::{ARROW_SYMBOL, UNIT_ATOM};
 use crate::metta::runner::stdlib::unit_result;
 use crate::metta::runner::str::{Str, ATOM_TYPE_STRING};
@@ -11,7 +11,7 @@ use std::rc::Rc;
 use crate::space::grounding::GroundingSpace;
 use crate::metta::text::SExprParser;
 use crate::metta::runner::{ModuleLoader, RunContext, DynSpace, Metta, MettaMod};
-use crate::atom::gnd::*;
+use hyperon_atom::gnd::*;
 use crate::metta::runner::number::{Number, ATOM_TYPE_NUMBER};
 
 pub static FILEIO_METTA: &'static str = include_str!("fileio.metta");
