+++ lib/src/metta/runner/stdlib/debug.rs
@@ -1,13 +1,13 @@
-use crate::*;
+use hyperon_atom::*;
 use crate::metta::*;
 use crate::metta::text::Tokenizer;
 use hyperon_common::collections::{SliceDisplay, Equality, DefaultEquality};
 use hyperon_common::assert::compare_vec_no_order;
-use crate::atom::matcher::atoms_are_equivalent;
+use hyperon_atom::matcher::atoms_are_equivalent;
 use crate::metta::runner::stdlib::{grounded_op, regex, unit_result};
 use crate::metta::runner::bool::*;
 use crate::metta::runner::str::*;
-use crate::atom::gnd::GroundedFunctionAtom;
+use hyperon_atom::gnd::GroundedFunctionAtom;
 
 use std::convert::TryInto;
