+++ lib/src/metta/runner/stdlib/package.rs
@@ -1,5 +1,5 @@
 use super::{grounded_op, regex, unit_result};
-use crate::*;
+use hyperon_atom::*;
 use crate::metta::*;
 use crate::metta::text::Tokenizer;
 use crate::metta::runner::{Metta, RunContext,
