+++ lib/src/metta/runner/stdlib/string.rs
@@ -1,4 +1,4 @@
-use crate::*;
+use hyperon_atom::*;
 use crate::metta::*;
 use crate::metta::text::Tokenizer;
 use crate::metta::runner::str::*;
