+++ lib/src/space/module.rs
@@ -91,7 +91,7 @@ impl SpaceMut for ModuleSpace {
 #[cfg(test)]
 mod test {
     use hyperon_common::assert_eq_no_order;
-    use crate::*;
+    use hyperon_atom::*;
     use crate::space::grounding::*;
     use super::*;
